feat: be - 소켓 서버에 세션 연결 #66
- express-session을 소켓 서버에 연결함
